+++ source/jormungandr/jormungandr/__init__.py
@@ -94,8 +94,8 @@ stat_manager = StatManager()
 
 from jormungandr.parking_space_availability.bss.bss_provider_manager import BssProviderManager
 bss_provider_manager = BssProviderManager(app.config['BSS_PROVIDER'])
-from jormungandr.parking_space_availability.car.car_parking_provider_manager import CarParkingProviderManager
-car_parking_provider_manager = CarParkingProviderManager(app.config['CAR_PARKING_PROVIDER'])
+from jormungandr.parking_space_availability.car.car_park_provider_manager import CarParkingProviderManager
+car_park_provider_manager = CarParkingProviderManager(app.config['CAR_PARK_PROVIDER'])
 
 
 from jormungandr import api
