+++ source/jormungandr/jormungandr/parking_space_availability/__init__.py
@@ -26,7 +26,7 @@
 # IRC #navitia on freenode
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
-from abstrcat_parking_places_provider import AbstractParkingPlacesProvider
+from abstract_parking_places_provider import AbstractParkingPlacesProvider
 from abstract_provider_manager import AbstractProviderManager
 from abstract_provider_manager import get_from_to_pois_of_journeys
 from bss.stands import Stands
