+++ source/jormungandr/jormungandr/parking_space_availability/abstract_parking_places_provider.py
