+++ source/jormungandr/jormungandr/parking_space_availability/car/tests/star_test.py
@@ -46,14 +46,14 @@ poi = {
     }
 }
 
-def car_parking_space_availability_start_support_poi_test():
+def car_park_space_availability_start_support_poi_test():
     """
     STAR car provider support
     """
     provider = StarProvider({'Keolis Rennes'}, 'toto', 42)
     assert provider.support_poi(poi)
 
-def car_parking_space_get_information_test():
+def car_park_space_get_information_test():
     parking_places = ParkingPlaces(available=4,
                                    occupied=3,
                                    available_PRM=2,
