+++ source/navitiacommon/navitiacommon/default_values.py
@@ -100,7 +100,7 @@ priority = 0
 bss_provider = True
 
 # activate / desactivate call to car parking provider
-car_parking_provider = True
+car_park_provider = True
 
 #Maximum number of connections allowed in journeys is calculated as
 #max_additional_connections + minimum connections among the journeys
