+++ source/navitiacommon/navitiacommon/models.py
@@ -314,7 +314,7 @@ class Instance(db.Model):
     bss_provider = db.Column(db.Boolean, default=default_values.bss_provider,
                                   nullable=False, server_default=true())
 
-    car_parking_provider = db.Column(db.Boolean, default=default_values.car_parking_provider,
+    car_park_provider = db.Column(db.Boolean, default=default_values.car_park_provider,
                                      nullable=False, server_default=true())
 
     max_additional_connections = db.Column(db.Integer, default=default_values.max_additional_connections,
